[Promotion][Core] Moved ActionFactory from Promotion to Core
