+++ src/Sylius/Component/Core/Factory/ActionFactoryInterface.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Sylius\Component\Promotion\Factory;
+namespace Sylius\Component\Core\Factory;
 
 use Sylius\Component\Promotion\Model\ActionInterface;
 use Sylius\Component\Resource\Factory\FactoryInterface;
